requirements.txt
weaviate
fastapi
uvicorn
weaviate-client
python-dotenv
langgraph
graph
langchain